requirements.txt
colorlog
nltk
transformers
torch
torchvision
salesforce-lavis
labml-nn
